# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BEYOND-MOBILE-BACKEND",
        "title": "BEYOND-MOBILE-BACKEND",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/users/auth/forgot-pin": {
            "post": {
                "description": "request access to renew pin on forgot moment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Forgot Pin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-SIGNATURE",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-DEVICE-ID",
                        "name": "X-DEVICE-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-TIMESTAMP",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LATITUDE",
                        "name": "X-LATITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LONGITUDE",
                        "name": "X-LONGITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Register Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_request.ForgotPinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.BasicSuccess"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.UserNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.CommonError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/auth/login": {
            "post": {
                "description": "Login using email or username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-SIGNATURE",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-DEVICE-ID",
                        "name": "X-DEVICE-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-TIMESTAMP",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LATITUDE",
                        "name": "X-LATITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LONGITUDE",
                        "name": "X-LONGITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Email Login Request",
                        "name": "loginEmailRequest",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_request.LoginEmailRequest"
                        }
                    },
                    {
                        "description": "Phone number Login Request",
                        "name": "loginPhoneNumberRequest",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_request.LoginPhoneNumberRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.LoginSuccess"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.DeferenceDevice"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.UserNotFound"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.CommonError"
                        }
                    }
                }
            }
        },
        "/api/v1/users/auth/logout": {
            "post": {
                "description": "user log-out from apps",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Log-out User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-SIGNATURE",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-DEVICE-ID",
                        "name": "X-DEVICE-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-TIMESTAMP",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LATITUDE",
                        "name": "X-LATITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LONGITUDE",
                        "name": "X-LONGITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Logout Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_request.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.BasicSuccess"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "backend-mobile-api_model_dto_request.ForgotPinRequest": {
            "type": "object",
            "required": [
                "callback_url",
                "device_id",
                "device_info",
                "email"
            ],
            "properties": {
                "callback_url": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                },
                "device_info": {
                    "$ref": "#/definitions/backend-mobile-api_model_dto_request.DeviceData"
                },
                "email": {
                    "type": "string"
                }
            }
        },
...
        "backend-mobile-api_model_dto_request.LoginEmailRequest": {
            "type": "object",
            "required": [
                "device_id",
                "email",
                "latitude",
                "longitude",
                "pin",
                "place",
                "type"
            ],
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "latitude": {
                    "type": "string"
                },
                "longitude": {
                    "type": "string"
                },
                "pin": {
                    "type": "string"
                },
                "place": {
                    "type": "string"
                },
                "type": {
                    "enum": [
                        "EMAIL",
                        "PHONE_NUMBER",
                        "BIOMETRIC"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/backend-mobile-api_model_enum.LoginType"
                        }
                    ],
                    "example": "EMAIL"
                }
            }
        },
...
        "backend-mobile-api_model_dto_swagger.BasicSuccess": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": ""
                },
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "string",
                    "example": "00"
                }
            }
        },
        "backend-mobile-api_model_dto_swagger.CommonError": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": "time out"
                },
                "message": {
                    "type": "string",
                    "example": "server is busy"
                },
                "status_code": {
                    "type": "string",
                    "example": "119"
                }
            }
        },
        "backend-mobile-api_model_dto_swagger.DeferenceDevice": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": ""
                },
                "message": {
                    "type": "string",
                    "example": "deference device"
                },
                "status_code": {
                    "type": "string",
                    "example": "129"
                }
            }
        },
...
        "backend-mobile-api_model_dto_swagger.UserNotFound": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": ""
                },
                "message": {
                    "type": "string",
                    "example": "user not found"
                },
                "status_code": {
                    "type": "string",
                    "example": "117"
                }
            }
        },
...
    }
}